move vmdebugger logic to its own class
